+++ src/components/Icon/Icon.tsx
@@ -11,6 +11,10 @@ import nicaragua from './img/nicaragua.svg';
 import pakistan from './img/pakistan.svg';
 import austria from './img/austria.svg';
 import spain from './img/spain.svg';
+import usa from './img/usa.svg';
+import france from './img/france.svg';
+import japan from './img/japan.svg';
+import ukraine from './img/ukraine.svg';
 import earth from './img/earth.svg';
 import verdaccio from './img/verdaccio.svg';
 import filebinary from './img/filebinary.svg';
@@ -23,11 +27,15 @@ export interface IconsMap {
   brazil: string;
   spain: string;
   china: string;
+  usa: string;
   nicaragua: string;
   pakistan: string;
   austria: string;
+  france: string;
   germany: string;
   india: string;
+  japan: string;
+  ukraine: string;
   earth: string;
   verdaccio: string;
   license: string;
@@ -54,6 +62,10 @@ export const Icons: IconsMap = {
   time,
   version,
   germany,
+  usa,
+  france,
+  japan,
+  ukraine,
 };
 
 export interface Props {